+++ packages/components/src/components/common/loading-message/loading-message-styles.ts
@@ -1,0 +1,8 @@
+import { type Styles } from 'jss';
+import { getHiddenTextJssStyle } from '../../../styles';
+
+export const getFunctionalComponentLoadingMessageStyles = (): Styles<'loading'> => {
+  return {
+    loading: getHiddenTextJssStyle(),
+  };
+};