+++ packages/components/src/components/switch/switch-styles.ts
@@ -13,6 +13,7 @@ import {
   prefersColorSchemeDarkMediaQuery,
 } from '../../styles';
 import { borderWidthBase, spacingStaticSmall, textSmallStyle } from '@porsche-design-system/utilities-v2';
+import { getFunctionalComponentLoadingMessageStyles } from '../common/loading-message/loading-message-styles';
 
 const getColors = (
   checked: boolean,
@@ -188,5 +189,7 @@ export const getComponentCss = (
         height: '28px',
       },
     }),
+    // .loading
+    ...getFunctionalComponentLoadingMessageStyles(),
   });
 };