+++ packages/components/src/components/switch/switch.tsx
@@ -1,4 +1,4 @@
-import { Component, Element, Event, type EventEmitter, h, Host, type JSX, Listen, Prop } from '@stencil/core';
+import { Component, Element, Event, type EventEmitter, h, Host, type JSX, Listen, Prop, Watch } from '@stencil/core';
 import type { BreakpointCustomizable, PropTypes, Theme } from '../../types';
 import {
   ALIGN_LABELS,
@@ -14,6 +14,7 @@ import {
 import { getComponentCss } from './switch-styles';
 import type { SwitchAlignLabel, SwitchAlignLabelDeprecated, SwitchUpdateEventDetail } from './switch-utils';
 import { getSwitchButtonAriaAttributes } from './switch-utils';
+import { LoadingMessage, loadingId } from '../common/loading-message/loading-message';
 
 const propTypes: PropTypes<typeof Switch> = {
   alignLabel: AllowedTypes.breakpoint<SwitchAlignLabel>(ALIGN_LABELS),
@@ -61,11 +62,29 @@ export class Switch {
   /** Emitted when checked status is changed. */
   @Event({ bubbles: false }) public update: EventEmitter<SwitchUpdateEventDetail>;
 
+  private initialLoading: boolean = false;
+
   @Listen('click', { capture: true })
   public onClick(e: MouseEvent): void {
     if (isDisabledOrLoading(this.disabled, this.loading)) {
       e.stopPropagation();
     }
+  }
+
+  @Watch('loading')
+  public loadingChanged(newVal: boolean): void {
+    if (newVal) {
+      // don't reset initialLoading to false
+      this.initialLoading = newVal;
+    }
+  }
+
+  public connectedCallback(): void {
+    this.initialLoading = this.loading;
+  }
+
+  public componentWillLoad(): void {
+    this.initialLoading = this.loading;
   }
 
   public componentShouldUpdate(newVal: unknown, oldVal: unknown): boolean {
@@ -110,23 +129,20 @@ export class Switch {
           type="button"
           role="switch"
           aria-labelledby="label" // only relevant for axe-core because of https://github.com/dequelabs/axe-core/issues/1393
+          aria-describedby={this.loading ? loadingId : undefined}
           onClick={this.onSwitchClick}
         >
           {/* it's necessary to always render toggle and a conditionally nested spinner, for smooth transitions */}
           <span class="toggle">
             {this.loading && (
-              <PrefixedTagNames.pSpinner
-                class="spinner"
-                size="inherit"
-                theme={this.theme}
-                aria={{ 'aria-label': 'Loading state' }}
-              />
+              <PrefixedTagNames.pSpinner class="spinner" size="inherit" theme={this.theme} aria-hidden="true" />
             )}
           </span>
         </button>
         <label id="label" htmlFor="switch">
           <slot />
         </label>
+        <LoadingMessage loading={this.loading} initialLoading={this.initialLoading} />
       </Host>
     );
   }